refactor #step1 MethodMatcher 인터페이스 구현
