+++ src/test/java/study/MethodMatcher.java
@@ -1,0 +1,8 @@
+package study;
+
+import java.lang.reflect.Method;
+
+public interface MethodMatcher {
+
+    boolean matches(Method method, Class targetClass, Object[] args);
+}
