+++ src/test/java/study/PrefixSayMethodMatcher.java
@@ -1,0 +1,13 @@
+package study;
+
+import java.lang.reflect.Method;
+
+public class PrefixSayMethodMatcher implements MethodMatcher {
+
+    public static final String PREFIX_METHOD = "say";
+
+    @Override
+    public boolean matches(Method method, Class targetClass, Object[] args) {
+        return method.getName().startsWith(PREFIX_METHOD);
+    }
+}
